+++ src/delegation/store/memory.rs
@@ -69,13 +69,13 @@ use web_time::SystemTime;
 /// linkStyle 1 stroke:orange;
 /// ```
 #[derive(Debug, Clone, PartialEq)]
-pub struct MemoryStore<DID: Did + Ord, V: varsig::Header<C>, C: Codec + TryFrom<u32> + Into<u32>> {
+pub struct MemoryStore<DID: Did + Ord, V: varsig::Header<C>, C: Codec + TryFrom<u64> + Into<u64>> {
     ucans: BTreeMap<Cid, Delegation<DID, V, C>>,
     index: BTreeMap<Option<DID>, BTreeMap<DID, BTreeSet<Cid>>>,
     revocations: BTreeSet<Cid>,
 }
 
-impl<DID: Did + Ord, V: varsig::Header<C>, C: Codec + TryFrom<u32> + Into<u32>> Default
+impl<DID: Did + Ord, V: varsig::Header<C>, C: Codec + TryFrom<u64> + Into<u64>> Default
     for MemoryStore<DID, V, C>
 {
     fn default() -> Self {
@@ -88,7 +88,7 @@ impl<DID: Did + Ord, V: varsig::Header<C>, C: Codec + TryFrom<u32> + Into<u32>> 
 }
 
 // FIXME check that UCAN is valid
-impl<DID: Did + Ord + Clone, V: varsig::Header<Enc>, Enc: Codec + TryFrom<u32> + Into<u32>>
+impl<DID: Did + Ord + Clone, V: varsig::Header<Enc>, Enc: Codec + TryFrom<u64> + Into<u64>>
     Store<DID, V, Enc> for MemoryStore<DID, V, Enc>
 {
     type DelegationStoreError = (); // FIXME misisng
